Migrate UI

+++ includes/level_edit/level_edit.h
@@ -1,29 +1,31 @@
+
 #ifndef LEVEL_EDIT_H
 #define LEVEL_EDIT_H
 
 #include "../level/level.h"
 
 typedef struct Position_s {
-    int x;
-    int y;
+    short x;
+    short y;
 } Position;
 
-int checkHitBox(Level* level, int x, int y, int xSize, int ySize);
+int checkHitBox(Level* level, short x, short y, short xSize, short ySize);
 
-int supprimerObjet(Level* level, int x, int y);
+int supprimerObjet(Level* level, short x, short y);
 
-int poserBlock(Level* level, int x, int y);
-int poserVie(Level* level, int x, int y);
-int poserBomb(Level* level, int x, int y);
-int poserTrap(Level* level, int x, int y);
-int poserGate(Level* level, int x, int y, int gateColor);
-int poserKey(Level* level, int x, int y, int keyColor);
-int poserDoor(Level* level, int x, int y, int doorNumber);
-int poserExit(Level* level, int x, int y);
-int poserStart(Level* level, int x, int y);
-int poserProbe(Level* level, int x, int y);
-int poserRobot(Level* level, int x, int y);
-int poserLadder(Level* level, int x, int y);
-int poserPlayer(Level* level, int x, int y);
+int poserBlock(Level* level, short x, short y);
+int poserVie(Level* level, short x, short y);
+int poserBomb(Level* level, short x, short y);
+int poserTrap(Level* level, short x, short y);
+int poserGate(Level* level, short x, short y, int8_t gateColor);
+int poserKey(Level* level, short x, short y, int8_t keyColor);
+int poserDoor(Level* level, short x, short y, int8_t doorNumber);
+int poserExit(Level* level, short x, short y);
+int poserStart(Level* level, short x, short y);
+int poserProbe(Level* level, short x, short y);
+int poserRobot(Level* level, short x, short y);
+int poserLadder(Level* level, short x, short y);
+int poserPlayer(Level* level, short x, short y);
 
 #endif
+
